Integrate PR#768 and add handlng for MacOS.

+++ hdl/quartus/build_bladerf.sh
@@ -296,7 +296,7 @@ fi
 nios_system=../fpga/ip/altera/nios_system
 
 # 9a484b436: Windows-specific workaround for Quartus bug
-if [ "x$(uname)" != "xLinux" ]; then
+if [ "x$(uname)" != "xLinux" -a "x$(uname)" != "xDarwin"]; then
     QUARTUS_BINDIR=$QUARTUS_ROOTDIR/bin
     export QUARTUS_BINDIR
     echo "## Non-Linux OS Detected (Windows?)"
@@ -482,4 +482,6 @@ printf "%s %02d:%02d:%02d\n" "Total Build Time:" "$(($SECONDS / 3600))" "$((($SE
 echo ""
 
 # Delete empty SOPC directories in the user's home directory
-find ~ -maxdepth 1 -type d -empty -iname "sopc_altera_pll*" -delete
+if [ "x$(uname)" == "xLinux" -o "x$(uname)" == "xDarwin" ]; then
+    find ~ -maxdepth 1 -type d -empty -iname "sopc_altera_pll*" -delete
+fi
